refactor: simplify architecture by removing hexagonal structure and flattening CLI app
