+++ budget_state.json
@@ -1,6 +1,6 @@
 {
-  "reserve": "700",
+  "reserve": "796.0",
   "available_funds": "0",
   "rent": "0",
-  "taxes": "0"
+  "taxes": "24.0"
 }
