--- core/exceptions.py
@@ -1,2 +1,0 @@
-class CancelledTransaction (Exception):
-    pass
