+++ transactions.jsonl
@@ -2,3 +2,4 @@
 {"id": 2, "timestamp": "12-07-2025 14:28:28", "type": "expense", "amount": "900", "comment": "test2", "category": "5", "tax_status": null}
 {"id": 3, "timestamp": "12-07-2025 14:28:39", "type": "expense", "amount": "55", "comment": "test3", "category": "4", "tax_status": null}
 {"id": 4, "timestamp": "12-07-2025 15:43:28", "type": "income", "amount": "700", "comment": "00", "category": null, "tax_status": "\u043d\u0435\u0442"}
+{"id": 5, "timestamp": "13-07-2025 16:57:55", "type": "income", "amount": "120", "comment": "333", "category": null, "tax_status": "\u0434\u0430"}
